Changed the way I traverse the bvh from bfs to dfs. That actually makes a big difference because of early discarding of BVH that are further compared to the ray distance

+++ RayTracer/src/Scene.h
@@ -65,7 +65,7 @@ struct Scene {
     void createUniqueVertices(const rapidobj::Mesh& mesh, const rapidobj::Attributes& attributes);
     uint32_t getIndexOfVertex(const Key& key, const rapidobj::Attributes& attributes, std::unordered_map<Key, int>& uniqueIndexKeys);
     std::unique_ptr<BVH> prepBvh(int left, int right, const AABB& box);
-    size_t Scene::splitWithSAH(const AABB& box, const glm::vec3& boxDimensions, glm::vec3& splitPointLeft, glm::vec3& splitPointRight, int left, int right, int axis);
+    int Scene::splitWithSAH(const AABB& box, const glm::vec3& boxDimensions, glm::vec3& splitPointLeft, glm::vec3& splitPointRight, int left, int right, int axis);
     //std::unique_ptr<BVH> Scene::prepBvh(int left, int right, int level);
 };
